Move StatsModelsLinearRegression
